+++ uikit/src/main/java/com/sendbird/uikit/fragments/ChannelFragment.java
@@ -1041,14 +1041,17 @@ public class ChannelFragment extends BaseGroupChannelFragment implements OnItemC
         DialogListItem save = new DialogListItem(R.string.sb_text_channel_anchor_save, R.drawable.icon_download);
         DialogListItem delete = new DialogListItem(R.string.sb_text_channel_anchor_delete, R.drawable.icon_delete);
 
+        DialogListItem retry = new DialogListItem(R.string.sb_text_channel_anchor_retry, 0);
+        DialogListItem deleteFailed = new DialogListItem(R.string.sb_text_channel_anchor_delete, 0);
+
         DialogListItem[] actions = null;
         BaseMessage.SendingStatus status = message.getSendingStatus();
         switch (type) {
             case VIEW_TYPE_USER_MESSAGE_ME:
                 if (status == BaseMessage.SendingStatus.SUCCEEDED) {
                     actions = new DialogListItem[]{copy, edit, delete};
-                } else if (status == BaseMessage.SendingStatus.FAILED || status == BaseMessage.SendingStatus.CANCELED) {
-                    actions = new DialogListItem[]{delete};
+                } else if (MessageUtils.isFailed(message)) {
+                    actions = new DialogListItem[]{retry, deleteFailed};
                 }
                 break;
             case VIEW_TYPE_USER_MESSAGE_OTHER:
@@ -1057,8 +1060,8 @@ public class ChannelFragment extends BaseGroupChannelFragment implements OnItemC
             case VIEW_TYPE_FILE_MESSAGE_VIDEO_ME:
             case VIEW_TYPE_FILE_MESSAGE_IMAGE_ME:
             case VIEW_TYPE_FILE_MESSAGE_ME:
-                if (status == BaseMessage.SendingStatus.FAILED || status == BaseMessage.SendingStatus.CANCELED) {
-                    actions = new DialogListItem[]{delete};
+                if (MessageUtils.isFailed(message)) {
+                    actions = new DialogListItem[]{retry, deleteFailed};
                 } else {
                     actions = new DialogListItem[]{delete, save};
                 }
@@ -1084,7 +1087,7 @@ public class ChannelFragment extends BaseGroupChannelFragment implements OnItemC
                     messageAnchorDialog.show();
                     anchorDialogShowing = true;
                 }
-            } else if (MessageUtils.isUnknownType(message)) {
+            } else if (MessageUtils.isUnknownType(message) || MessageUtils.isFailed(message)) {
                 if (getContext() == null || getFragmentManager() == null) return;
                 DialogUtils
                         .buildItemsBottom(actions, createMessageActionListener(message))
@@ -1175,7 +1178,12 @@ public class ChannelFragment extends BaseGroupChannelFragment implements OnItemC
             } else if (key == R.string.sb_text_channel_anchor_edit) {
                 editMessage(message);
             } else if (key == R.string.sb_text_channel_anchor_delete) {
-                showWarningDialog(message);
+                if (MessageUtils.isFailed(message)) {
+                    Logger.dev("delete");
+                    deleteMessage(message);
+                } else {
+                    showWarningDialog(message);
+                }
             } else if (key == R.string.sb_text_channel_anchor_save) {
                 if (Build.VERSION.SDK_INT > Build.VERSION_CODES.P) {
                     download((FileMessage) message);
@@ -1193,6 +1201,8 @@ public class ChannelFragment extends BaseGroupChannelFragment implements OnItemC
                         }
                     });
                 }
+            } else if (key == R.string.sb_text_channel_anchor_retry) {
+                resendMessage(message);
             }
         };
     }